backend/test_pdf_fix.py
#!/usr/bin/env python3
"""
Test PDF extraction with the fixed pdfminer implementation
"""
import asyncio
import sys
import os
from pathlib import Path
from io import BytesIO

# Add current directory to path
sys.path.append(os.path.dirname(os.path.abspath(__file__)))

from app.services.document_processor import document_processor
import logging

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)

async def test_pdf_extraction_methods():
    """Test each PDF extraction method individually"""
    print("🧪 Testing PDF extraction methods...")
    
    # Check if we have any PDF files
    uploads_dir = Path("./uploads")
    pdf_files = list(uploads_dir.glob("*.pdf")) if uploads_dir.exists() else []
    
    if not pdf_files:
        print("❌ No PDF files found in uploads directory")
        print("📝 Creating a simple test to verify the fix...")
        
        # Test the pdfminer function directly
        try:
            from pdfminer.high_level import extract_text as pdfminer_extract_text
            from pdfminer.layout import LAParams
            print("✅ pdfminer imports working correctly")
            
            # Test the LAParams configuration
            laparams = LAParams(
                char_margin=2.0,
                line_margin=0.5,
                word_margin=0.1,
                boxes_flow=0.5,
                detect_vertical=True,
                all_texts=False
            )
            print("✅ LAParams configuration working correctly")
            
            # Test that we're calling extract_text correctly (without the problematic output_type parameter)
            print("✅ Fixed pdfminer call should work without 'output_type' parameter")
            return True
            
        except Exception as e:
            print(f"❌ pdfminer setup failed: {e}")
            return False
    
    # Test with actual PDF file
    test_file = pdf_files[0]
    print(f"🧪 Testing with actual PDF: {test_file.name}")
    
    try:
        with open(test_file, 'rb') as f:
            file_content = f.read()
        
        # Test pdfminer specifically
        pdf_file = BytesIO(file_content)
        text_pdfminer = document_processor._extract_with_pdfminer(pdf_file)
        
        if text_pdfminer:
            print(f"✅ pdfminer extraction successful: {len(text_pdfminer)} characters")
        else:
            print("⚠️ pdfminer extraction returned empty text (might be normal for some PDFs)")
        
        # Test the overall PDF extraction
        pdf_file.seek(0)
        extracted_text = document_processor._extract_text_from_pdf(file_content)
        
        if extracted_text:
            print(f"✅ Overall PDF extraction successful: {len(extracted_text)} characters")
            print(f"📝 First 200 characters: {extracted_text[:200]}...")
            return True
        else:
            print("❌ Overall PDF extraction failed")
            return False
            
    except Exception as e:
        print(f"❌ PDF extraction test failed: {e}")
        import traceback
        traceback.print_exc()
        return False

if __name__ == "__main__":
    print("🚀 PDF Extraction Fix Test")
    print("=" * 50)
    
    success = asyncio.run(test_pdf_extraction_methods())
    
    if success:
        print("\n✅ PDF extraction fix verified!")
    else:
        print("\n❌ PDF extraction fix failed!")
