backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pinecone-client==3.2.2
pymongo==4.6.0
google-generativeai==0.3.2
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
pymupdf==1.23.14
pdfminer.six==20231228
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
python-dotenv==1.0.0
motor==3.3.2
numpy==1.24.3
aiofiles==23.2.1
httpx==0.25.2
certifi